+++ tests/fixest_tests.R
@@ -1204,6 +1204,16 @@ est_feols = feols(y ~ x | grp + tm, data=d)
 test(se(est_feols, se = "st")["x"], se(est_lm)["x"])
 
 #
+# Heteroskedasticity-robust
+#
+
+se_white_lm_HC1 = sqrt(vcovHC(est_lm, type = "HC1")["x", "x"])
+se_white_lm_HC0 = sqrt(vcovHC(est_lm, type = "HC0")["x", "x"])
+
+test(se(est_feols, se = "hetero"), se_white_lm_HC1)
+test(se(est_feols, se = "hetero", ssc = ssc(adj = FALSE, cluster.adj = FALSE)), se_white_lm_HC0)
+
+#
 # Clustered
 #
 
@@ -1215,15 +1225,6 @@ se_CL_grp_lm_HC0 = sqrt(vcovCL(est_lm, cluster = d$grp, type = "HC0")["x", "x"])
 test(se(est_feols, ssc = ssc(fixef.K = "full")), se_CL_grp_lm_HC1)
 test(se(est_feols, ssc = ssc(adj = FALSE, fixef.K = "full")), se_CL_grp_lm_HC0)
 
-#
-# Heteroskedasticity-robust
-#
-
-se_white_lm_HC1 = sqrt(vcovHC(est_lm, type = "HC1")["x", "x"])
-se_white_lm_HC0 = sqrt(vcovHC(est_lm, type = "HC0")["x", "x"])
-
-test(se(est_feols, se = "hetero"), se_white_lm_HC1)
-test(se(est_feols, se = "hetero", ssc = ssc(adj = FALSE, cluster.adj = FALSE)), se_white_lm_HC0)
 
 #
 # Two way
@@ -1235,13 +1236,81 @@ se_CL_2w_feols = se(est_feols, se = "twoway")
 
 test(se(est_feols, se = "twoway", ssc = ssc(fixef.K = "full", cluster.df = "conv")), se_CL_2w_lm)
 
+# 
+# HC2/HC3
+#
+
+# HC2/HC3
+base = iris
+base$w = runif(nrow(base))
+
+est_feols = feols(Sepal.Length ~ Sepal.Width | Species, base)
+est_lm = lm(Sepal.Length ~ Sepal.Width + factor(Species), base)
+
+test(
+  vcov(est_feols, "hc2", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC2")["Sepal.Width", "Sepal.Width"]
+)
+
+test(
+  vcov(est_feols, "hc3", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC3")["Sepal.Width", "Sepal.Width"]
+)
+
+# HC2/HC3 with weights
+base = iris
+base$w = runif(nrow(base))
+
+est_feols = feols(Sepal.Length ~ Sepal.Width | Species, base, weights = ~ w)
+est_lm = lm(Sepal.Length ~ Sepal.Width + factor(Species), base, weights = base$w)
+
+test(
+  vcov(est_feols, "hc2", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC2")["Sepal.Width", "Sepal.Width"]
+)
+
+test(
+  vcov(est_feols, "hc3", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC3")["Sepal.Width", "Sepal.Width"]
+)
+
+# HC2/HC3 with GLM
+base$Sepal.Length = floor(base$Sepal.Length)
+est_feglm = feglm(
+  Sepal.Length ~ Sepal.Width | Species, base, 
+  "poisson", weights = ~ w
+)
+est_glm = glm(
+  Sepal.Length ~ Sepal.Width + factor(Species), base, 
+  family = poisson(), weights = base$w
+)
+
+test(
+  vcov(est_feglm, "hc2", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_glm, "HC2")["Sepal.Width", "Sepal.Width"]
+)
+test(
+  vcov(est_feglm, "hc3", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_glm, "HC3")["Sepal.Width", "Sepal.Width"]
+)
+
+# Fail when P_ii = 1
+base$Species = as.character(base$Species)
+base[1, "Species"] = "foo"
+
+est_pii_singular = feols(Sepal.Length ~ Sepal.Width | Species, base)
+
+test(
+  vcov(est_pii_singular, "hc2"), "err"
+)
+
 #
 # Checking the calls work properly
 #
 
 data(trade)
 
-est_pois = femlm(Euros ~ log(dist_km)|Origin+Destination, trade)
+est_pois = femlm(Euros ~ log(dist_km) | Origin + Destination, trade)
 
 se_clust = se(est_pois, se = "cluster", cluster = "Product")
 test(se(est_pois, cluster = trade$Product), se_clust)
@@ -1822,6 +1891,11 @@ glm_probit  = glm(y_bin ~ x1 + x2, family = binomial("probit"), base)
 feglm_probit = feglm(y_bin ~ x1 + x2, base, binomial("probit"))
 test(hatvalues(feglm_probit), hatvalues(glm_probit))
 
+# Fixed effects
+fm_fe  = lm(y ~ x1 + x2 + factor(species), base)
+ffm_fe = feols(y ~ x1 + x2 | species, base)
+test(hatvalues(ffm_fe), hatvalues(fm_fe))
+
 
 ####
 #### sandwich ####
@@ -2316,6 +2390,142 @@ test(unlist(coef(est_split)), unlist(coef(est_split_noup)))
 est_fe = update(est_mult, . ~ . | fe)
 est_fe_noup = feols(c(y1, y2) ~ x1 | fe, base_mult)
 test(unlist(coef(est_fe)), unlist(coef(est_fe_noup)))
+
+####
+#### sparse_model_matrix ####
+####
+
+base = iris
+names(base) = c("y1", "x1", "x2", "x3", "species")
+base$y2 = 10 + rnorm(150) + 0.5 * base$x1
+base$x4 = rnorm(150) + 0.5 * base$y1
+base$fe2 = rep(letters[1:15], 10)
+base$fe2[50:51] = NA
+base$y2[base$fe2 == "a" & !is.na(base$fe2)] = 0
+base$x2[1:5] = NA
+base$x3[6] = NA
+base$fe3 = rep(letters[1:10], 15)
+base$id = rep(1:15, each = 10)
+base$time = rep(1:10, 15)
+
+base_bis = base[1:50, ]
+base_bis$id = rep(1:5, each = 10)
+base_bis$time = rep(1:10, 5)
+
+
+res = feols(y1 ~ x1 + x2 + x3, base)
+sm1 = sparse_model_matrix(
+  res, type = "lhs",
+  na.rm = TRUE
+)
+test(length(sm1), res$nobs)
+
+sm1_na = sparse_model_matrix(
+  res, data = base, 
+  type = "lhs",
+  na.rm = FALSE
+)
+test(length(sm1_na), res$nobs_origin)
+test(max(abs(sm1_na - base$y1), na.rm = TRUE), 0)
+
+
+sm2 = sparse_model_matrix(
+  res, type = c("lhs", "rhs"), data = base, 
+  combine = FALSE, na.rm = FALSE
+)
+y = sm2[["lhs"]]
+X = sm2[["rhs"]]
+obs_rm = res$obs_selection$obsRemoved
+res_bis = solve(
+  crossprod(as.matrix(X)[obs_rm, ]), 
+  crossprod(as.matrix(X)[obs_rm, ], as.matrix(y)[obs_rm, ])
+)
+test(as.numeric(res_bis), res$coefficients)
+
+# No constant
+res_nocons = feols(mpg ~ 0 + i(cyl), mtcars)
+sm_nocons = sparse_model_matrix(res_nocons, type = "rhs")
+test("(Intercept)" %in% colnames(sm_nocons), FALSE)
+
+# Lag 
+res_lag = feols(y1 ~ l(x1, 1:2) + x2 + x3, base, panel = ~id + time)
+sm_lag = sparse_model_matrix(res_lag, type = "rhs")
+test(nrow(sm_lag), nobs(res_lag))
+
+
+# TODO: Fix poly and newdata
+# With poly
+# res_poly = feols(y1 ~ poly(x1, 2), base)
+# works
+# res_poly = feols(y1 ~ x1, base)
+# sm_poly_old = sparse_model_matrix(res_poly)
+# sm_poly_new = sparse_model_matrix(res_poly, data = base_bis)
+# test(sm_poly_old[1:50, 3], sm_poly_new[, 3])
+
+
+# Interacted fixef
+res = feols(y1 ~ x1 + x2 + x3 | species^fe2, base)
+sm_ife = sparse_model_matrix(res, data = base, type = "fixef", collin.rm = FALSE)
+
+# fixef
+res = feols(y1 ~ x1 + x2 + x3 | species + fe2, base)
+sm_fe = sparse_model_matrix(res, data = base, type = "fixef")
+test(ncol(sm_fe), 17)
+
+sm_fe_no_collin_rm = sparse_model_matrix(res, data = base, type = "fixef", collin.rm = FALSE)
+test(ncol(sm_fe_no_collin_rm), 18)
+
+sm_fe_base_bis = sparse_model_matrix(res, data = base_bis, type = "fixef")
+
+
+# Time-varying slopes
+res_slopes = feols(y1 ~ x1 + x2 + x3 | fe2[I(x2+1)], data = base[7:48, ])
+sm_slopes = sparse_model_matrix(res_slopes, type = "fixef")
+
+# IV
+res_iv = feols(y1 ~ x1 | x2 ~ x3, base)
+
+sm_rhs1 = sparse_model_matrix(res_iv, type = "iv.rhs1")
+test(colnames(sm_rhs1)[-1], c("x3", "x1"))
+
+sm_rhs2 = sparse_model_matrix(res_iv, type = "iv.rhs2")
+test(colnames(sm_rhs2)[-1], c("fit_x2", "x1"))
+
+sm_endo = sparse_model_matrix(res_iv, type = "iv.endo")
+test(colnames(sm_endo), "x2")
+
+sm_exo  = sparse_model_matrix(res_iv, type = "iv.exo")
+test(colnames(sm_exo)[-1], "x1")
+
+sm_inst  = sparse_model_matrix(res_iv, type = "iv.inst")
+test(colnames(sm_inst), "x3")
+
+# several
+res_mult = feols(y1 ~ x1 | species | x2 ~ x3, base)
+
+sm_lhs_rhs_fixef = sparse_model_matrix(res_mult, type = c("lhs", "iv.rhs2", "fixef"))
+test(colnames(sm_lhs_rhs_fixef), c("y1", "fit_x2", "x1", "species::setosa", "species::versicolor", "species::virginica"))
+
+
+# non-linear model
+res_pois = fepois(Sepal.Length ~ Sepal.Width + Petal.Length | Species, iris)
+sp_pois = sparse_model_matrix(res_pois, data = iris)
+
+
+# make sure names are correct
+test_col_names <- function(est) {
+  m <- fixest::sparse_model_matrix(est, collin.rm = FALSE)
+  q <- test(all(names(coef(est)) %in% colnames(m)), TRUE)
+  return(invisible(NULL))
+}
+test_col_names(feols(mpg ~ i(am) + disp | vs, mtcars))
+test_col_names(feols(mpg ~ i(am, i.cyl) + disp | vs, mtcars))
+test_col_names(feols(mpg ~ i(am, hp) + disp | vs, mtcars))
+test_col_names(feols(mpg ~ i(am):hp + disp | vs, mtcars))
+test_col_names(feols(mpg ~ factor(am) + disp | vs, mtcars))
+test_col_names(feols(mpg ~ factor(am):factor(cyl) + disp | vs, mtcars, notes = FALSE))
+test_col_names(feols(mpg ~ factor(am):hp + disp | vs, mtcars))
+test_col_names(feols(mpg ~ poly(hp, degree = 2) + disp | vs, mtcars))
 
 ####
 #### fitstat ####